+++ src/app/lib/analyticsUtils/sendComponentTrackingBeacon/index.test.ts
@@ -1,5 +1,5 @@
-import * as sendBeacon from '../../../lib/analyticsUtils/sendBeacon';
-import * as analyticsUtils from '../../../lib/analyticsUtils';
+import * as sendBeacon from '../sendBeacon';
+import * as analyticsUtils from '..';
 import { sendEventBeacon } from '.';
 
 const sendBeaconSpy = jest.spyOn(sendBeacon, 'default');
