+++ src/app/lib/analyticsUtils/types.ts
@@ -1,6 +1,6 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import { ComponentTrackingProps } from '#app/components/PageViewTracking/types';
 import { EventTrackingMetadata } from '#app/models/types/eventTracking';
+import { Platforms, Services } from '#app/models/types/global';
 import {
   CLICK_EVENT,
   STATIC_ATI_VIEW_TRACKING,
@@ -8,6 +8,26 @@ import {
 } from './analytics.const';
 
 export type ATIEventType = typeof VIEW_EVENT | typeof CLICK_EVENT;
+
+export interface ComponentTrackingProps {
+  campaignID?: string;
+  componentName?: string;
+  format?: string;
+  pageIdentifier?: string;
+  platform?: Platforms;
+  producerId?: string;
+  producerName?: string;
+  service?: Services;
+  statsDestination?: string;
+  type?: string;
+  advertiserID?: string;
+  url?: string;
+  detailedPlacement?: string;
+  useReverb?: boolean;
+  experimentVariant?: string;
+  ampExperimentName?: string;
+  preventNavigation?: string;
+}
 
 export type EventTrackingData = EventTrackingMetadata & ComponentTrackingProps;
